+++ Entities/EnemyProjectile.cpp
@@ -1,7 +1,7 @@
 #include "EnemyProjectile.h"
 
 EnemyProjectile::EnemyProjectile(std::shared_ptr<Character> enemy, Vector2D<double> relStartPos, std::shared_ptr<Character> player, 
-    double xVel, double yVel, int colliderWidth, int colliderHeight, int damage, const char* fileName)
+    double xVel, double yVel, double colliderWidth, double colliderHeight, int damage, const char* fileName)
     : Projectile(fileName, enemy->getPos().getx() + relStartPos.getx(), enemy->getPos().gety() + relStartPos.gety(), xVel, yVel, colliderWidth, colliderHeight, damage)
 {
     m_player = player;
